Refactor users/accounts library and implement UPGs
This is based on the Debian documentation for UPGs: each user's primary
group is a group of the same name and ID of which they are the sole
member, and umask is set to 002 rather than 022 so that created files
are group-writeable by default. This means shared-writeable directories
can easily be created just by chgrping them and enabling setgid.

Doing this was more aggravating than it should have been, because nixos
doesn't expose the necessary configuration settings to add new PAM
modules to the PAM configuration, and pam_umask is needed to change the
umask settings. Ultimately I implemented this by wiggling some
undocumented, internal-only settings of the PAM nixos module.